Factor out OpenQASM 3 frontend from api.cpp (#19)
Move the driver code that constitutes the OpenQASM 3 frontend from
api.cpp's `compile()` into `lib/Frontend/OpenQASM3/`. Move all code
specific to the OpenQASM 3 frontend into the namespace
`qssc::frontend::openqasm3`. Move all command line parameters specific
to the OpenQASM 3 frontend there (to show up as a separate category in
the output `qss-compiler --help`).

This change is purely a move of existing code, in preparation of
upcoming changes. That is, structural changes such as introducing a
unified frontend interface are deliberately out of scope of this PR.

Co-authored-by: Thomas Alexander <thomasalexander2718@gmail.com>