Get version from CMake root project's Git repo. (#26)
Currently, the Python package versioning mechanism is a bit of a mess.
This is because we (IBM) package this Python module after baking our
proprietary targets into it, and then version it to match the version
of our drivers. So that means the version number we use comes from
outside this project.

The way this works currently is we use setuptools_scm to get the Git
tag version from the repo found at CMAKE_SOURCE_DIR, which is
the directory containing the outter-most CMakeLists.txt. This way, we
get the version from the qss-compiler repo itself if it is the root
CMake project, or from the the parent project's repo if it's being
used as a Git submodule.

However, this doesn't work when QSS compiler is consumed as a
submodule that isn't located at the root of the parent repo (e.g.
a monorepo).

To work around this, we now use CMake to invoke Git to determine the
root CMake project's Git repo root, determine the version of that using
SCM, and thread it down to the Python package.

Note
----
None of this is the right thing to do. In the long-term, we should
avoid baking targets into the Python module in favor of dynamic loading
and discovery. And then, we should start versioning QSS compiler as
a target-agnostic package with its own versioning scheme, and
version dynamically loaded targets as separate Python packages.