Enable CMAKE_INSTALL_RPATH_USE_LINK_PATH option on Apple.
The CMAKE_INSTALL_RPATH_USE_LINK_PATH is used to ensure the paths to
all dynamic libraries is included in the final executable. We are
moving to prefer dylib over static libraries on MacOS, and this option
allows the final executable to be run without requiring the user to
set the location of all dynamic libraries in their DYLD_LIBRARY_PATH
environment variable.