Clear cache in main if deps changed.
This is necessary otherwise old versions of Conan packages will stick
around in the cache, and it'll grow unbounded over time. We already use
7/10 GB, so this would be a problem. In fact, it'd be an immediate issue
whenever we rev LLVM.

In practice, we get most of our dependencies from Conan center, so this
is really only an inconvenience in that we need to rebuild LLVM whenever
we rev any Conan package after a merge to main.