Fix conversion QUIR/STD -> LLVM (#42)
To translate operations from a dialect to LLVMIR, the dialect must have
a registered implementation of
`LLVMTranslationDialectInterface`. Due to changes during the migration
to LLVM-14, the `mock_target` implementation of `QUIRToStandard` was not
updated to accommodate the new interface. When testing the generation of
a QuantumExecutionModule (QEM) from a QASM 3.0 input, the following
error was observed:

```
error: cannot be converted to LLVM IR: missing `LLVMTranslationDialectInterface`
registration for dialect for op: builtin.unrealized_conversion_cast
```

ultimately, producing no controller binary for the target. Due to the
lack of test coverage for `mock_target`, the functionality for emitting
a QEM output from QASM 3 input was broken for some time.

This PR fixes the above issue, adding the necessary pieces for QUIR to
have a registered implementation of
`LLVMTranslationDialectInterface`, as well as additional testing to
check for correct output when emitting a QEM from QASM.

Closes https://github.com/Qiskit/qss-compiler/issues/41.