Introduce new `PayloadV2` class using Cap'n Proto schema
The qss-compiler presently outputs a libzip-based (archive) for
downstream consumption. An improved strategy would be to instead
have a standardized schema to present to clients consuming
qss-compiler output.

This commit brings in the first iteration using a Cap'n Proto
schema as a standard interface for qss-compiler output. Note this
is not the end of the story, but aims to serve as the scaffolding
for compliant interfacing between layers of the software stack.