Conditionally link LLVM AArch64, anchor CI Python minor version (#54)
Originally, #42 introduced a link requirement on `LLVMAArch64CodeGen`,
which is not built into our LLVM build unless the host is macOS. I tried
to fix this in #45 by linking using the CMake `LLVM::LLVM` target
instead, which contains only the libs actually present in the LLVM
build. However, it seems that the ordering of these libraries isn't
correct (an issue with our LLVM Conan recipe), so this causes other
linking failures.

To fix this instead, we use CMake to detect if the AArch64 target is
present in LLVM, and remove the corresponding libs from our link line if
not. This works because we can keep the proper order with respect to the
other LLVM libs we need to link in, and filter out what we don't need.

Separately, we lock the Python minor version for CI.