Switch to Conan CMakeToolchain and CMakeDeps. (#60)
Previously, we used the `cmake_find_package` generator for dependencies
and the `cmake` generator to propagate Conan build settings into the
CMake build system.

In Conan 2.0, the only supported generators are
[CMakeDeps](https://docs.conan.io/2/reference/tools/cmake/cmakedeps.html#conan-tools-cmakedeps)
and
[CMakeToolchain](https://docs.conan.io/2/reference/tools/cmake/cmaketoolchain.html#conan-tools-cmaketoolchain),
which we switch to here. Note that we have not yet switched to Conan
2.0, but these changes get us most of the way there.

The most notable developer impact is that builds done directly with
CMake will require that the generated toolchain be specified:

```
cmake -G Ninja -DCMAKE_TOOLCHAIN_FILE=conan_toolchain.cmake ..
```

Builds done via `conan build` should be unchanged.