Refactor `TargetRegistry` to a templatized `PluginRegistry` and concrete `TargetSystemRegistry` (#63)
Together with #62, it substitutes most of #58

Here I'm adding a new `PluginRegistry<PluginInfo>` class to implement a
conventional registry-factory for `PluginInfo` items. On top of this
registry I'm implementing a `TargetSystemRegistry` that creates not only
`TargetSystemInfo` instances, but `TargetSystem` ones with some
cache-like functionality based on `mlir::MLIRContext` input.

This refactor supports a future `PayloadRegistry` as a specialization of
`PluginRegistry`.

All existing functionality is preserved.