Introduce OpenQASM 3 (OQ3) dialect (#34)
We introduce the first phase of decoupling our OpenQASM 3 frontend from
our QUIR middle-end to define the structure and space for growing our
OpenQASM 3 frontend with the means of MLIR.

This is the first phase, migrating only the relevant QUIR `Operation`s
to OQ3, leaving QUIR types and attributes as is. Additionally,
pattern-matching can be further, more cleanly, separated between QUIR
and OQ3, as well as lowering between the two dialects (presently, QUIR
and OQ3 go directly to standard dialects).