Add standardized configuration handling (#83)
This PR adds standardized configuration handling for "global"
configuration through the `QSSConfig` and `QSSConfigBuilder`s which
populate the configuration. To begin with a small number of
configuration parameters such as
`targetName` and `targetConfigurationPath` have been defined. Support
for a single `CLIConfigBuilder` has been added
which seamlessly replaces some existing cli arguments to instead
populate the `QSSConfig` rather than being used directly. It is possible
to set the `QSSConfig` for a `MLIRContext`. (this is being done by
registry similar to the current class target. I'm still not convinced
this is a better approach than inheriting from the `MLIRContext` for a
`QSSContext` where these may be stored directly.)

This PR will be followed up with a PR implementing a `QSSConfigBuilder`
for environment variables.


The `API.cpp` class has been refactored to use these (and while doing so
a significant amount of general refactoring was performed).