Fix M1 CI Failure in VariableEliminationPass  (#103)
This PR makes a small change to the VariableEliminationPass to fix CI
failures on Apple M1s. See issue #102.

The code in question iterates over the users of a the result of the
operation to delete the users. On M1 it is necessary to copy the users
to a SmallVector and then iterate as erasing a user appears to change
the iteration over getUsers().