Delay inside if block fix (#129)
This PR fixes an error introduced with the circuits PR merger. 

The QUIRGen circuits functions use a different builder to track
insertion in the parent block, which was not being updated when control
flow constructs were encountered. This PR updates the circuits builder
tracking to included updates for such control flow changes.