Potential tablegen dependency fix (#139)
This PR attempts to fix a missing dependency issue in the cmake system. 

Occasionally we will run into build errors of the form:
```
No such file or directory
   28 | #include "Dialect/QUIR/IR/QUIRTypes.h.inc"
      |          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
```

This is caused by a missing dependency between the files created by
tablegen and any source file that attempts to include something that
includes them (such as the dialect headers).

I have added explicit dependencies to the `mlir-headers` cmake target
that is defined by `add_mlir_dialect` to hopefully resolve this problem.
Unfortunately source file creation is not a typical part of the build
process outside of LLVM, so it's hard to be sure that cmake will still
respect this dependency the way we want.