Improvement to MergeCircuit and ReorderMeasurements for Parameters (#157)
Adds additional improvements for handling input parameters in
the MergeCircuit pass and the ReorderMeasurement pass.

1. The passes will now move more parameters related operations out of
the way when attempting to reorder measurements and merge circuits.

2. The `MergeCircuitsPass::mergeCallCircuits` has been simplified.

3. This PR also adds the `--include-source` command line options as a
debug flag. If this flag is passed when the direct input method is also
used then the input source will be stored in the output payload at
`manifest\input.{qasm|mlir}` depending on the file format.