Configure build mode for LIT tests.
This change allows us to query the build mode used to build the compiler in LIT tests.
This will enable the REQUIRES constraint to disable tests when the compiler is not
built with asserts or in debug mode. See https://llvm.org/docs/TestingGuide.html\#constraining-test-execution
for more details on constraining test execution.