Add a target compilation manager framework (#200)
Adds a target compilation manager (in analogy to the pass manager) to
the qss-compiler. This manager provides a systematic compilation of
trees of target systems and their children. It also provides a threaded
implementation of the target compilation manager that uses the shared
threadpool from the MLIRContext to provide simultaneous threading of
both target passes and across target compilation with unified resource
management through the shared threadpool.