Add support for MLIR bytecode through the qe-compile (#250)
This PR adds support for MLIR bytecode to the qe-compiler for both
inputting, outputting and through the Python bindings. It was
necessarily to refactor the compiler's main API entrypoints to align
with MLIROptMain (just in a general a more flexible API) to support
making the necessary Python binding calls.

This PR also removes the need for `--direct` as an option in favour of
supporting more typical piping directly to the tool.