Add test for sequence before accessing argument (#261)
Add a guard to verify that a sequence is present in the
argumentToDuration map before accessing the argument number. The
sequence may not be present if a playOp is present in a sequence which
is not actually called in a program. This may happen when manually
editing a MLIR input file for testing (for example commenting out a
pulse.call_sequence).