Add CircuitBarrierCircuitPattern to MergeCircuits (#239)
Adds a CircuitBarrierCircuitPattern to the MergeCircuits Pass to allow
a quir.barrier in-between two quir.circuits to be merged into a circuit 
quir.call_circuit with the merged circuit containing the barrier(s). 
Adding this pass enables more efficient pulse scheduling.

The overall pass has been refactored to factor out a number of helper
functions.